warning police

+++ src/util/types.cpp
@@ -81,10 +81,10 @@ int64_t bigDivide(int64_t A, int64_t B, int64_t C)
 
 bool iequals(const std::string& a, const std::string& b)
 {
-    unsigned int sz = a.size();
+    size_t sz = a.size();
     if(b.size() != sz)
         return false;
-    for(unsigned int i = 0; i < sz; ++i)
+    for(size_t i = 0; i < sz; ++i)
         if(tolower(a[i]) != tolower(b[i]))
             return false;
     return true;
